RecBivariateSpline changed to 1 degree (kx=1,ky=1)
This helps to smooth out interpolation results.  Previously, no smooth
values were observed in the grid because the spline interpolation had
too many degrees.

Also - previous version did not interpolate if the nearest neighbour
(to the .NGH file pt being interpolated) was equal to 0m. This caused
strange artefacts in the depth grid.